Organization of the main file, more clear and clean
