+++ controller/plr_controller.py
@@ -1,5 +1,9 @@
 import logic.polynomial_regression as pr
 import time
+import numpy as np
+import matplotlib.pyplot as plt
+from utils.generador_csv import polynomial_regression_csv
+from utils.read_csv import read_csv
 
 
 def polynomial_regression(x, y, degree):
@@ -21,3 +25,26 @@ def polynomial_regression(x, y, degree):
     time_taken = end_time - start_time
 
     return w, b, error, time_taken
+
+
+def polynomial_regression_func():
+    degree = 3
+    polynomial_regression_csv(degree)
+    x, y = read_csv('data/polynomial_regression_data.csv')
+    w, b, error, time_taken = polynomial_regression(x, y, degree)
+    print(f"Coefficients (w): {w}")
+    print(f"Intercept (b): {b}")
+    print(f"Error: {error}, Time taken: {time_taken:.4f} seconds")
+
+    line = [sum(w[i] * (xi ** (degree - i))
+                for i in range(len(w))) + b for xi in x]
+
+    x_sorted, line_sorted = zip(*sorted(zip(x, line)))
+
+    plt.scatter(x, y, label='Data Points')
+    plt.plot(x_sorted, line_sorted, color='orange', label='Polynomial Fit')
+    plt.xlabel('X')
+    plt.ylabel('Y')
+    plt.title('Polynomial Regression')
+    plt.legend()
+    plt.show()
