+++ data/linear_regression.csv
@@ -1,51 +1,51 @@
 x,y
-66.51097017441995,109.5821002769947
-62.54013266422794,100.29458875538468
-73.26279437109363,96.33688625449292
--5.989756087460854,53.677385976419224
-4.944114735657379,78.2079250134267
-71.6709937523199,52.29278686214289
-62.81920978612347,14.37890168799098
-23.50051191030536,9.941682068271092
-63.77820327366243,14.782191549640935
-94.6229258104299,21.05238260495688
-19.558759843941495,118.74958605787498
--31.536310532631294,56.15137411506777
-96.82946048344613,127.00719223044524
-6.125798481844802,54.652153077656784
-41.17072116471005,64.21115758733528
-67.67179147102838,179.4205231258344
-28.59244679263439,73.17941871877929
-48.634195072804786,53.60430397721718
-93.51831239434091,59.347398109237176
-95.46739227909151,-0.07399973933229465
-58.52579385800098,63.114308587289024
-43.517065740911136,54.49378650813566
-29.52184736796429,37.256804839327444
-50.41304575752643,-36.66703174260991
-64.79704373477358,41.09531120454095
-136.16043361573062,-31.687602181767957
--37.59951323941158,113.09837406331616
-28.08158739289847,40.511242033467454
--19.01784148454148,82.82431744595198
-49.83095036147117,76.45555056519575
-80.55324495969661,14.641210833719306
-54.517890799449106,36.7265438861459
--12.602661588088267,103.46708944144879
-49.71166889789906,33.62781274041076
-143.21741783673502,115.94122383442524
-115.30199165404622,44.18284019096836
-71.23119611937088,78.67572634235873
-10.261377147971668,43.44647839365375
-65.8461992821373,70.51116375308204
-49.48038228975476,60.589852804367965
-99.2131297933296,104.55626547704738
-22.8128820201524,10.895860770933492
-24.913776883418393,56.25848303878399
-41.41214578057208,67.89744756679556
-26.696872716242616,66.80076114740265
-72.2781128580886,49.004879969894645
--5.124563368582379,63.467574940814195
-39.51708923618925,12.453535201805828
--62.084233007463254,124.58187042188048
-120.86891178814267,14.359333234939065
+72.94381172188656,43.64025169408966
+50.356133324078925,3.3323451491105214
+76.16480489849597,14.757085855635278
+63.763029521930704,130.3968083027204
+52.09172090697599,32.53476743112011
+-2.272258868182604,100.93414784365062
+-29.40607502525679,31.664990971029688
+46.16254053349388,58.38603521068039
+66.17243775061806,86.01872435283984
+108.11565750789667,61.507194193643436
+91.84431048194492,105.70738460850558
+17.87941883623317,106.5651888388603
+123.63569616366519,16.384127295482934
+23.55572951720804,79.80021082667784
+20.118995718857505,81.39488474481576
+3.8104469678801056,67.6550904600117
+15.552188226785344,112.64859244894794
+31.68442195062195,87.70442119685971
+55.656344007531104,63.282055216783505
+-20.593442658230572,118.2715938739991
+23.625795072182875,-13.851669995055122
+33.36581256254596,72.11480673973129
+42.35637091232181,66.55234353679626
+40.16584056702786,16.332285116337616
+56.82720178396732,57.61065147977036
+-1.3766967255917453,67.42385350178259
+-5.844424122616729,-7.7721779081111215
+76.93300062811804,86.92538683732559
+64.50314766153853,138.37548363707498
+11.760333990765702,23.194486168933697
+32.24508691201225,80.23977436483234
+-17.834070034000444,25.7685905573587
+66.59745900522125,38.934981606187804
+87.34033377059404,64.19388041433186
+3.90419780790742,64.4193107128094
+63.12618142068226,49.10758927384299
+22.804169917849066,38.29243357941637
+29.698491318765615,57.842655843931716
+151.0926235493631,121.29984538149614
+112.37832590262452,9.14770617364097
+39.27555482736002,-1.967953909226651
+-21.30217718374225,53.71013508404034
+79.78437870979843,-41.15814393709384
+52.705920659950735,86.7732329258096
+51.40605877199536,0.3886490558373694
+55.509301215246396,86.81022343525014
+5.284961287046597,63.63957486014016
+128.34264434460482,-25.398434874048572
+69.9711509308092,20.97968169645585
+-9.233069579908225,11.023338643937379
