+++ data/multiple_linear_regression_data.csv
@@ -1,51 +1,51 @@
 x1,x2,y
-87.3036926362387,88.03266568898252,5530.393765785249
--9.187622150858747,16.994670569630923,17615.76959773506
-32.93086256297767,29.633156577962296,10315.166502398191
-96.71972221049751,22.55035697550509,15308.997571961165
-29.36382965701475,64.92748796418483,-830.9281044065774
--7.774154271190966,53.1271770442661,11945.025166409716
--2.076859432461493,38.35837329380424,11022.20083871731
--7.2127216189973336,-16.82311099506745,-8956.802857523546
--35.510590742401746,40.31794578732719,25054.167303258604
-37.32275990447991,45.6456859374393,-7812.360201434088
-57.18807675288387,80.391529381374,13866.250674436524
-80.35349517221225,102.79947146301099,17609.17903516585
--32.665344666721204,6.5350228581953616,18980.18210985016
-114.7370257845285,64.4253284349361,47456.30908020142
-74.2436010516352,95.2425293647039,-712.6855036992935
-33.48371154150165,58.924186720584736,-685.9265994660514
-79.79702783402206,35.50912249290626,14087.054515161357
-45.64405320544537,48.39629543521601,8919.760845851166
-59.66724746933724,-36.05576130640123,16191.461803269845
-73.91871866081408,12.861596193682857,23668.47441388904
-81.32305512338905,84.399724705543,21702.892278478506
-28.94877434720511,22.291359653297086,10139.2175924936
-87.80406243846515,92.48722560682701,31358.6130529652
-45.707541077520496,119.25937443131583,6007.890961256481
-7.814817354207598,55.49392923794693,314.18540423130526
-32.1188939101807,1.1168541344222156,1291.7005632551359
-32.2711989220466,35.00048770562759,-9570.449989621626
-94.31929034649502,87.62940608541491,34802.99840733115
-49.58293105745068,60.62836832696868,7132.0352364295095
-119.39980820516053,39.4818322831364,-3863.149573353097
-73.7979088972562,107.29039900211384,3025.7285279868534
-55.568104486007094,39.65135509968401,35401.52776265821
-66.15011165537052,42.158895714956444,8474.885507802892
--9.362670687642364,31.374240489048873,288.2972402443466
-70.9093064016937,85.5702315602989,7029.451961323696
-129.58332201328213,-19.55823443551485,6932.824008178418
-27.303281100255568,5.695988849915231,4223.084948658024
-54.18120998982063,104.68522743514396,9866.8236134201
-105.10215913175517,-0.3681150728122624,6515.084813586801
-40.36402720868711,13.186501831885291,12696.743734661679
-49.91347808601443,92.84910015930836,23427.957654471244
-68.947515970148,26.988382629368832,3354.8891690079645
-53.80481208700063,-4.140069689092908,-4243.6238230739045
-53.38595968016293,23.090046951306057,18425.576757627394
-70.60580425446582,107.3048317382956,16642.984837104512
-85.14612859919319,53.13807976541672,257.1958242184073
--6.169454010563314,-52.308884756711294,-1734.440128297325
-114.98295383314374,51.51676417477208,-2146.883761715999
-6.231384262055952,90.73338545856744,-6759.475875689738
-63.369585619210255,79.59100912236433,7546.461413587411
+87.613649578969,23.342928273049367,16031.266643421222
+86.6567218992983,87.89461758450209,20937.02100179771
+13.186732503360833,14.395921577576805,2559.000089487435
+65.52889870956142,12.672768734765228,43662.92973733617
+122.98122883799977,13.500293487551945,-20709.02366168882
+27.77471004742404,49.376807547356925,8058.048409246918
+132.24005233635734,-9.80565383518831,40856.01098717851
+89.91296333046418,59.82542687979462,6251.554015980069
+115.91722543416266,56.02738940280868,1808.3690540636271
+49.2896195429083,-13.249319529296884,-634.3485869928018
+48.342173343859656,68.98507822634936,8952.341205465884
+-14.642022383550792,75.65558727320447,3979.669996219493
+165.30417092575516,11.246276749074514,34974.13356861981
+14.653699973133243,58.61629751317872,-2067.5747551213412
+21.55174177325546,-16.032361938094027,15522.329690629442
+-12.641909901611726,56.730483176071765,10228.204376412425
+87.38681945679971,66.8694019291151,8389.274174947093
+105.61183718123866,33.72848622063283,20516.86929329088
+36.6154673901226,69.01120351831727,1582.1167486338236
+-5.7511479453711445,46.18597618857366,-5848.786314304785
+80.51577519527446,59.16762131807869,6206.914116349118
+76.586759509331,134.45480507786905,-6949.988727929849
+60.54460922984866,63.551639405371986,4685.662531746884
+55.66695492435449,17.53369492124284,19034.09282374167
+-16.765618874215164,9.855391127721713,3410.7998115943046
+0.8328703771599208,63.26573494873141,17200.0727747896
+64.13578466742264,113.1091513445801,-20374.59692328658
+70.35809626304645,-53.456000305046516,14692.178925554834
+60.91586056756944,13.696735446433296,14868.278513405703
+-8.696843301587624,64.33949471421958,17926.894504435684
+14.178659995695,143.88362301028053,5719.6138926863205
+124.9116167086131,-44.17606495059405,18324.14622306232
+72.36947948092583,64.47268600795613,19163.317167327867
+67.49805899790368,46.761245502473116,28242.545955490034
+18.002085262745073,138.48044420761727,19881.403387250644
+41.828433439631915,56.23765025791989,23623.857585780628
+50.85770121479939,59.700856101701255,-5933.7761519349515
+9.760059240741239,37.75028107623986,18756.594796341986
+49.555974416258195,-1.5021101725694272,9484.54287994398
+-7.660473246864122,73.84892270295335,-6980.289094292592
+83.19732366322188,82.7670246616043,29282.457868466932
+96.29659011711215,124.48914639179768,6186.015038230078
+121.3035701994144,29.189269732579323,-1981.7907102448112
+73.55297503671503,80.3822736482698,1844.723849555582
+35.013902024670685,36.111874656940195,12320.894524091098
+38.28201532037984,37.30408488418971,-1662.1232446858226
+24.365966594544254,8.778300467352288,19099.418726963784
+87.59347435042518,89.12723903517012,19838.380327302693
+22.803927944350164,-13.86382631981867,20524.6100609794
+69.94736237451573,12.07980583014654,8385.966575892397
